build: initial commit
This commit introduces Maven Wrapper to the project and includes a pre-commit hook. The Maven Wrapper facilitates building the project with the correct Maven version, while the pre-commit hook will apply code formatting before each commit. Also, initial pom.xml files for the parent project and the empty-spring-boot module, along with a readme file explaining the purpose of the project, were created.